requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
pytest
jupyter
